[PATCH 15/17] Get rid of the dynamic_control_mode ControlMode
This is a left over from before the dynamic_interface_module() rewrite,
not only is it no longer needed its code is plain wrong, as it still
contains a prototype for the old way different dynamic_interface_module()
and tries to call it as if it were the old version, if that code path where
to ever be invoked bad things (tm) would happen. Luckily it is never used,
so simply nuke it.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
 interface/Makefile.am |    6 --
 interface/dynamic_c.c |  130 -------------------------------------------------
 timidity/controls.c   |   10 ----
 timidity/timidity.c   |   31 +++---------
 4 files changed, 7 insertions(+), 170 deletions(-)
 delete mode 100644 interface/dynamic_c.c